MI5-69 Odvojen panel za pretraživanje korisnika
U AccountSearchViewu, odvojena je komponenta za search panel u zasebni file. Uz to, izrađene su male preinake u strukturi komponenti.
